Extract the reputation to network-peerset
